+++ src/main/groovy/geometry/Vector3.groovy
@@ -4,16 +4,16 @@ import groovy.transform.Immutable
 
 @Immutable
 class Vector3 {
-    double x = 0
-    double y = 0
-    double z = 0
+    float x = 0
+    float y = 0
+    float z = 0
 
     Vector3 cross(Vector3 v) {
-        return new Vector3(y * v.z - z * v.y, z * v.x - x * v.z, x * v.y - y * v.x);
+        return new Vector3((float)y * v.z - z * v.y,(float) z * v.x - x * v.z,(float) x * v.y - y * v.x);
     }
 
     Vector3 plus( Vector3  v)  {
-        return new Vector3(x + v.x, y + v.y, z + v.z)
+        return new Vector3((float)x + v.x,(float) y + v.y,(float) z + v.z)
     }
 
     Vector3 negative()  {
@@ -21,27 +21,27 @@ class Vector3 {
     }
 
     Vector3 multiply(double v)  {
-        return new Vector3(x * v, y * v, z * v);
+        return new Vector3((float)x * v, (float)y * v, (float)z * v);
     }
 
-    double multiply( Vector3  v)  {
+    float multiply( Vector3  v)  {
         return x * v.x + y * v.y + z * v.z;
     }
 
     Vector3 minus( Vector3  v)  {
-        return new Vector3(x - v.x, y - v.y, z - v.z);
+        return new Vector3((float)x - v.x,(float) y - v.y,(float) z - v.z);
     }
     //	Vector3 operator-(){
     //		return Vector3(-x, -y, -z);
     //	}
-    Vector3 div(double s)  {
-        return new Vector3(x / s, y / s, z / s);
+    Vector3 div(float s)  {
+        return new Vector3((float)x / s,(float) y / s,(float) z / s);
     }
-    double lenght()  {
+    float lenght()  {
         return Math.sqrt(x * x + y * y + z * z);
     }
     Vector3 normalize()  {
-        double l = lenght()
+        float l = lenght()
         if (l > 0.0001)
             return this / l
         else
@@ -52,17 +52,17 @@ class Vector3 {
      * mutiplies this vector vith a 3x3 matrix in a 1D array, assumes matrix is 9 long,
      * returns a new vector which is the result
      */
-    Vector3 multMatrix(double [] m)  {
+    Vector3 multMatrix(float [] m)  {
         return new Vector3(//
-                x * m[0] + y * m[1] + z * m[2],//
-                x * m[3] + y * m[4] + z * m[5],//
-                x * m[6] + y * m[7] + z * m[8]//
+                (float)x * m[0] + y * m[1] + z * m[2],//
+                (float)x * m[3] + y * m[4] + z * m[5],//
+                (float) x * m[6] + y * m[7] + z * m[8]//
         );
     }
 
-    Vector3 rotate( Vector3  axis, double angle)  {
-        double s = Math.sin(angle);
-        double c = Math.cos(angle);
+    Vector3 rotate( Vector3  axis, float angle)  {
+        float s = Math.sin(angle);
+        float c = Math.cos(angle);
         Vector3 u = axis.normalize();
 
         def m = [
